feat(chat-widget): split AI section out of general/behavior settings
Move agent picker, knowledge base, and featured articles into a new
dedicated AI tab. Drop chat-jwt re-verification on /pusher/* — the
form-encoded auth body breaks JWT parsing and the passport already
gates identity at session start.
